Fixed removeDuplicates
